functions and nested loops

+++ 0x02-functions_nested_loops/main.h
@@ -12,4 +12,4 @@ int add(int, int);
 void print_to_98(int n);
 void print_times_table(int n);
 void natural(void);
-int fibonacci(int );
+long fibonacci(long );
